pyproject.toml
[tool.poetry]
name = "記事抜粋アプリ"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "app" },
]

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.23.2"
langserve = {extras = ["server"], version = ">=0.0.30"}
pydantic = "<2"
langchain-core = "^0.2.5"
langchain-community = "^0.2.5"
langchain-openai = "^0.1.8"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
pytest-playwright = "^0.5.1"

[tool.poetry.group.dev.dependencies]
langchain-cli = ">=0.0.15"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"